Adding the ability to opt-out from specific group (Previous way removed contact from all groups)

+++ app/controllers/incoming_message_controller.rb
@@ -5,15 +5,22 @@ class IncomingMessageController < ApplicationController
     number = params["From"]
     number  = Phony.normalize(number) unless number.nil?
 
-    if params["SmsSid"] && params[:Body] == "STOP"
+    if params["SmsSid"] && params[:Body].include?("STOP")
+      group_id = params[:Body].delete("STOP")
+      
       contact = Sms.find_by_normalized_entry(number)
+      group_contact = GroupContact.find_by_contact_id_and_group_id( contact.id, group_id)
+      group_contact.destroy unless group_contact.nil?
     elsif params["CallSid"]
+      #Need to figure out how to get input from incoming phone numbers
       contact = Phone.find_by_normalized_entry(number)
+      contact.destroy unless contact.nil?
+      # group_contact = GroupContact.find_by_contact_id_and_group_id( contact.id, group_id)
+      # group_contact.destroy
+
     end
     if contact.nil?
       render(:file => File.join(Rails.root, 'public/404'), :status => 404, :layout => false, content_type: "text/xml" ) and return
-    else
-      contact.destroy
     end
     render layout: false
   end
